requirements.txt
feedparser
pyyaml
click
fuzzywuzzy[speedup]
python-dateutil
python-telegram-bot
python-dotenv
openai 